Added Comparison Property board. Made program fully aspect ratio independent. Removed Crash Error due to pre-testing.

+++ include/blocks.h
@@ -44,7 +44,7 @@ struct ThermalBlocks {
 
 class BlockManager {
 public:
-	float GetTempDif(double Tempa, double Tempb);
+	double GetTempDif(double Tempa, double Tempb);
 	ThermalBlocks CreateBlock(double temp, SDL_FRect& rect, double A, double d, double mass, double emissivety, double specific_heat_energy, double k);
 	void renderallblocks(ThermalBlocks& blocks, SDL_Renderer* renderer);
 	void getHeldState(SDL_FRect& mouse, ThermalBlocks* block);
