+++ include/physics.h
@@ -5,10 +5,10 @@
 class Physics {
 
 public:
-	void DeepSpaceHeatTransfer(std::vector <ThermalBlocks> * block, float dt, double sigma, double emissivety);
+	void DeepSpaceHeatTransfer(std::vector <ThermalBlocks> * block, double dt, double sigma, double emissivety);
 	//float HeatTransfer();
-	void AddConduction(const ThermalBlocks& BlockA, const ThermalBlocks& BlockB, size_t& ba, size_t& bb, float dt, float transferspeed, std::vector<double>& tempsToadd, BlockManager& blockUser);
-	void AddRadiation(const ThermalBlocks& BlockA, const ThermalBlocks& BlockB, size_t& ba, size_t& bb, float dt, double sigma, std::vector<double>& tempsToadd, BlockManager& blockUser);
+	void AddConduction(const ThermalBlocks& BlockA, const ThermalBlocks& BlockB, size_t& ba, size_t& bb, double dt, float transferspeed, std::vector<double>& tempsToadd, BlockManager& blockUser);
+	void AddRadiation(const ThermalBlocks& BlockA, const ThermalBlocks& BlockB, size_t& ba, size_t& bb, double dt, double sigma, std::vector<double>& tempsToadd, BlockManager& blockUser);
 	void GetRGB(ThermalBlocks* blocks);
 	void getCoolingConstant(double sigma, std::vector<ThermalBlocks>& blocksinuse);
 	float setdt(std::vector<ThermalBlocks>& block, double sigma, BlockManager& blockUser);
